+++ src/components/manage/Widgets/ObjectWidget.jsx
@@ -5,9 +5,12 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Tab } from 'semantic-ui-react';
+import { Accordion, Segment } from 'semantic-ui-react';
+import { Field, Icon } from '@plone/volto/components';
+import AnimateHeight from 'react-animate-height';
 
-import Field from '@plone/volto/components/manage/Form/Field';
+import upSVG from '@plone/volto/icons/up-key.svg';
+import downSVG from '@plone/volto/icons/down-key.svg';
 
 /**
  * Renders a field set. Passes some of the values in the schema to the Field
@@ -25,7 +28,6 @@ import Field from '@plone/volto/components/manage/Form/Field';
 const FieldSet = ({
   block,
   data,
-  index,
   schema,
   value,
   errors,
@@ -73,29 +75,13 @@ const ObjectWidget = ({
   id,
   ...props
 }) => {
-  const createTab = React.useCallback(
-    (fieldset, index) => {
-      return {
-        menuItem: fieldset.title,
-        render: () => (
-          <Tab.Pane>
-            <FieldSet
-              block={block}
-              data={fieldset}
-              index={index}
-              schema={schema}
-              errors={errors}
-              value={value}
-              onChange={onChange}
-              id={id}
-            />
-          </Tab.Pane>
-        ),
-      };
-    },
-    [block, errors, id, onChange, schema, value],
-  );
+  const [currentActiveFieldset, setCurrentActiveFieldset] = React.useState(0);
+  function handleCurrentActiveFieldset(e, blockProps) {
+    const { index } = blockProps;
+    const newIndex = currentActiveFieldset === index ? -1 : index;
 
+    setCurrentActiveFieldset(newIndex);
+  }
   return schema.fieldsets.length === 1 ? (
     <>
       <FieldSet
@@ -110,7 +96,44 @@ const ObjectWidget = ({
       />
     </>
   ) : (
-    <Tab panes={schema.fieldsets.map(createTab)} /> // lazy loading
+    schema.fieldsets.map((fieldset, index) => (
+      <Accordion fluid styled className="form" key={fieldset.id}>
+        <div key={fieldset.id} id={`blockform-fieldset-${fieldset.id}`}>
+          <Accordion.Title
+            active={currentActiveFieldset === index}
+            index={index}
+            onClick={handleCurrentActiveFieldset}
+          >
+            {fieldset.title && <>{fieldset.title}</>}
+            {currentActiveFieldset === index ? (
+              <Icon name={upSVG} size="20px" />
+            ) : (
+              <Icon name={downSVG} size="20px" />
+            )}
+          </Accordion.Title>
+          <Accordion.Content active={currentActiveFieldset === index}>
+            <AnimateHeight
+              animateOpacity
+              duration={500}
+              height={currentActiveFieldset === index ? 'auto' : 0}
+            >
+              <Segment className="attached">
+                <FieldSet
+                  block={block}
+                  data={fieldset}
+                  index={index}
+                  schema={schema}
+                  errors={errors}
+                  value={value}
+                  onChange={onChange}
+                  id={id}
+                />
+              </Segment>
+            </AnimateHeight>
+          </Accordion.Content>
+        </div>
+      </Accordion>
+    ))
   );
 };